refactor(bmc-mock): move carbide-specific code to machine-a-tron
