+++ crates/bmc-mock/src/hw/bluefield3.rs
@@ -18,10 +18,7 @@
 use std::borrow::Cow;
 use std::sync::Arc;
 
-use carbide_utils::arch::CpuArchitecture;
 use mac_address::MacAddress;
-use rpc::machine_discovery::{BlockDevice, CpuInfo, DiscoveryInfo, DmiData, DpuData};
-use rpc::{NetworkInterface, PciDeviceProperties};
 use serde_json::json;
 
 use crate::{BootOptionKind, Callbacks, LogService, LogServices, hw, redfish};
@@ -269,84 +266,6 @@ impl Bluefield3<'_> {
             part_number: Some(format!("{}       ", self.part_number()).into()),
             firmware_version: Some(self.firmware_versions.dpu_nic.clone().into()),
             is_mat_dpu: true,
-        }
-    }
-
-    pub fn discovery_info(&self) -> DiscoveryInfo {
-        DiscoveryInfo {
-            network_interfaces: vec![],
-            infiniband_interfaces: vec![],
-            cpu_info: vec![CpuInfo {
-                model: "Cortex-A78AE".into(),
-                vendor: "ARM".into(),
-                sockets: 1,
-                cores: 16,
-                threads: 16,
-            }],
-            block_devices: std::iter::once(BlockDevice {
-                model: "KBG40ZPZ128G TOSHIBA MEMORY".into(),
-                revision: "AEGA0103".into(),
-                serial: "FAKESERNUM0".into(),
-                device_type: "disk".into(),
-            })
-            .chain((0..3).map(|_| BlockDevice {
-                model: "NO_MODEL".into(),
-                revision: "NO_REVISION".into(),
-                serial: "NO_SERIAL".into(),
-                device_type: "disk".into(),
-            }))
-            .collect(),
-            machine_type: CpuArchitecture::Aarch64.to_string(),
-            machine_arch: Some(rpc::utils::cpu_architecture_to_rpc(
-                CpuArchitecture::Aarch64,
-            )),
-            nvme_devices: vec![],
-            dmi_data: Some(DmiData {
-                board_name: "Bluefield-3 DPU".into(),
-                board_version: "AG".into(),
-                bios_version: "4.13.0-26-g337fea6bfd".into(),
-                bios_date: "Nov  3 2025".into(),
-                product_serial: self.product_serial_number.to_string(),
-                board_serial: "Unspecified Base Board Serial Number".into(),
-                chassis_serial: "Unspecified Chassis Board Serial Number".into(),
-                product_name: "BlueField-3 DPU".into(),
-                sys_vendor: "Nvidia".into(),
-            }),
-            dpu_info: Some(DpuData {
-                part_number: self.part_number().into(),
-                part_description: format!("NVIDIA Bluefield-3 {}", self.part_number()),
-                product_version: self.firmware_versions.dpu_nic.clone(),
-                factory_mac_address: self.base_mac().to_string(),
-                firmware_version: self.firmware_versions.dpu_nic.clone(),
-                firmware_date: "11.11.2025".into(),
-                switches: vec![],
-            }),
-            gpus: vec![],
-            memory_devices: vec![],
-            tpm_ek_certificate: None,
-            tpm_description: None,
-            ..Default::default()
-        }
-    }
-
-    pub fn host_nic_discovery_info(
-        &self,
-        path: &str,
-        slot: &str,
-        numa_node: i32,
-    ) -> NetworkInterface {
-        NetworkInterface {
-            mac_address: self.host_mac_address.to_string(),
-            pci_properties: Some(PciDeviceProperties {
-                vendor: "Mellanox Technologies".into(),
-                device: "MT43244 BlueField-3 integrated ConnectX-7 network controller".into(),
-                path: path.into(),
-                numa_node,
-                description: Some(
-                    "MT43244 BlueField-3 integrated ConnectX-7 network controller".into(),
-                ),
-                slot: Some(slot.into()),
-            }),
         }
     }
