+++ crates/bmc-mock/src/hw/bluefield4.rs
@@ -18,9 +18,7 @@
 use std::borrow::Cow;
 use std::sync::Arc;
 
-use carbide_utils::arch::CpuArchitecture;
 use mac_address::MacAddress;
-use rpc::machine_discovery::{DiscoveryInfo, DmiData, DpuData};
 use serde_json::json;
 
 use crate::{Callbacks, LogService, LogServices, hw, redfish};
@@ -246,31 +244,6 @@ impl Bluefield4<'_> {
                 firmware_version: None,
                 is_mat_dpu: true,
             },
-        }
-    }
-
-    pub fn discovery_info(&self) -> DiscoveryInfo {
-        DiscoveryInfo {
-            machine_type: CpuArchitecture::Aarch64.to_string(),
-            machine_arch: Some(rpc::utils::cpu_architecture_to_rpc(
-                CpuArchitecture::Aarch64,
-            )),
-            dmi_data: Some(DmiData {
-                board_name: "BlueField-4 DPU".into(),
-                product_serial: self.product_serial_number.to_string(),
-                board_serial: carbide_utils::DEFAULT_DPU_DMI_BOARD_SERIAL_NUMBER.into(),
-                chassis_serial: carbide_utils::DEFAULT_DPU_DMI_CHASSIS_SERIAL_NUMBER.into(),
-                product_name: "BlueField-4 DPU".into(),
-                sys_vendor: "Nvidia".into(),
-                ..Default::default()
-            }),
-            dpu_info: Some(DpuData {
-                part_number: self.part_number().into(),
-                part_description: format!("NVIDIA BlueField-4 {}", self.part_number()),
-                factory_mac_address: self.host_mac_address.to_string(),
-                ..Default::default()
-            }),
-            ..Default::default()
         }
     }
