+++ crates/bmc-mock/src/hw/dell_poweredge_r760_bf4.rs
@@ -18,10 +18,7 @@
 use std::borrow::Cow;
 use std::sync::Arc;
 
-use bmc_vendor::BMCVendor;
-use carbide_utils::arch::CpuArchitecture;
 use mac_address::MacAddress;
-use rpc::machine_discovery::{BlockDevice, CpuInfo, DiscoveryInfo, DmiData, MemoryDevice};
 use serde_json::json;
 
 use crate::{BootOptionKind, Callbacks, hw, redfish};
@@ -223,56 +220,6 @@ impl DellPowerEdgeR760Bf4<'_> {
     pub fn update_service_config(&self) -> redfish::update_service::UpdateServiceConfig {
         redfish::update_service::UpdateServiceConfig {
             firmware_inventory: vec![],
-        }
-    }
-
-    pub fn discovery_info(&self) -> DiscoveryInfo {
-        DiscoveryInfo {
-            network_interfaces: vec![self.bf4.discovery_info(BF4_SLOT)],
-            infiniband_interfaces: vec![],
-            cpu_info: vec![CpuInfo {
-                model: "Intel(R) Xeon(R) Gold 6354 CPU @ 3.00GHz".into(),
-                vendor: "GenuineIntel".into(),
-                sockets: 2,
-                cores: 18,
-                threads: 36,
-            }],
-            block_devices: (0..2)
-                .map(|n| BlockDevice {
-                    model: "Dell Ent NVMe v2 AGN RI U.2 1.92TB".into(),
-                    revision: "2.3.0".into(),
-                    serial: format!("FAKESERNUM{n}"),
-                    device_type: "".into(),
-                })
-                .collect(),
-            machine_type: CpuArchitecture::X86_64.to_string(),
-            machine_arch: Some(rpc::utils::cpu_architecture_to_rpc(CpuArchitecture::X86_64)),
-            nvme_devices: vec![],
-            dmi_data: Some(DmiData {
-                board_name: "01J4WF".into(),
-                board_version: "A05".into(),
-                bios_version: "2.2.7".into(),
-                bios_date: "12/19/2023".into(),
-                product_serial: self.product_serial_number.to_string(),
-                board_serial: format!(".{}.FAKESERNUM2.", self.product_serial_number),
-                chassis_serial: self.product_serial_number.to_string(),
-                product_name: "PowerEdge R760".into(),
-                // Logic of machine state handler depends on BMC
-                // vendor that is calculated from dmi_data.sys_vendor
-                // value.
-                sys_vendor: hw::bmc_vendor_to_udev_dmi(BMCVendor::Dell).into(),
-            }),
-            dpu_info: None,
-            gpus: vec![],
-            memory_devices: (0..8)
-                .map(|_| MemoryDevice {
-                    size_mb: Some(16384),
-                    mem_type: Some("DDR4".into()),
-                })
-                .collect(),
-            tpm_ek_certificate: None,
-            tpm_description: None,
-            ..Default::default()
         }
     }
 }
