+++ crates/bmc-mock/src/hw/generic_ami.rs
@@ -18,7 +18,6 @@
 use std::borrow::Cow;
 use std::sync::Arc;
 
-use rpc::DiscoveryInfo;
 use serde_json::json;
 
 use crate::{BootOptionKind, Callbacks, hw, redfish};
@@ -138,9 +137,5 @@ impl GenericAmi<'_> {
         redfish::update_service::UpdateServiceConfig {
             firmware_inventory: vec![],
         }
-    }
-
-    pub fn discovery_info(&self) -> DiscoveryInfo {
-        DiscoveryInfo::default()
     }
 }
