+++ crates/bmc-mock/src/hw/nic.rs
@@ -18,7 +18,6 @@
 use std::borrow::Cow;
 
 use mac_address::MacAddress;
-use rpc::machine_discovery::{NetworkInterface, PciDeviceProperties};
 pub type SlotNumber = usize;
 
 pub struct Nic<'a> {
@@ -44,30 +43,6 @@ impl Nic<'_> {
             firmware_version: None,
             mac_address: mac,
             is_mat_dpu: false,
-        }
-    }
-
-    pub fn discovery_info(&self, slot: SlotNumber) -> NetworkInterface {
-        let device_name = format!("enp{}s{}np0", slot >> 16, slot & 0xff);
-        let slot = format!("{:04x}:{:02x}:00.0", (slot >> 16), (slot & 0xFF));
-        NetworkInterface {
-            mac_address: self.mac_address.to_string(),
-            pci_properties: Some(PciDeviceProperties {
-                vendor: self
-                    .manufacturer
-                    .as_ref()
-                    .unwrap_or(&Cow::Borrowed(""))
-                    .to_string(),
-                device: self
-                    .model
-                    .as_ref()
-                    .unwrap_or(&Cow::Borrowed(""))
-                    .to_string(),
-                path: format!("/devices/pci0000:00/0000:00:00.0/{slot}/net/{device_name}"),
-                numa_node: 0,
-                description: self.description.as_ref().map(|v| v.to_string()),
-                slot: Some(slot),
-            }),
         }
     }
 }
