+++ crates/bmc-mock/src/hw/nvidia_dgx_h100.rs
@@ -18,10 +18,7 @@
 use std::borrow::Cow;
 use std::sync::Arc;
 
-use carbide_utils::arch::CpuArchitecture;
 use mac_address::MacAddress;
-use rpc::machine_discovery::{CpuInfo, Gpu, InfinibandInterface, MemoryDevice};
-use rpc::{BlockDevice, DiscoveryInfo, DmiData, NetworkInterface, NvmeDevice, PciDeviceProperties};
 use serde_json::json;
 
 use crate::json::JsonExt;
@@ -360,168 +357,6 @@ impl NvidiaDgxH100<'_> {
             })
             .collect(),
         }
-    }
-
-    pub fn discovery_info(&self) -> DiscoveryInfo {
-        DiscoveryInfo {
-            network_interfaces: self.discovery_info_network_interfaces(),
-            infiniband_interfaces: self.discovery_info_ib_interfaces(),
-            cpu_info: vec![CpuInfo {
-                model: "Intel(R) Xeon(R) Platinum 8480CL".into(),
-                vendor: "GenuineIntel".into(),
-                sockets: 2,
-                cores: 56,
-                threads: 112,
-            }],
-            block_devices: (0..2)
-                .map(|n| BlockDevice {
-                    model: "Micron_7450_MTFDKBG1T9TFR".into(),
-                    revision: "E2MU200".into(),
-                    serial: format!("MicronFAKESERNUM{n}"),
-                    device_type: "disk".into(),
-                })
-                .chain((0..8).map(|n| BlockDevice {
-                    model: "KCM6DRUL3T84".into(),
-                    revision: "0107".into(),
-                    serial: format!("KCMFAKESERNUM{n}"),
-                    device_type: "disk".into(),
-                }))
-                .collect(),
-            machine_type: CpuArchitecture::X86_64.to_string(),
-            machine_arch: Some(rpc::utils::cpu_architecture_to_rpc(CpuArchitecture::X86_64)),
-            nvme_devices: (0..2)
-                .map(|n| NvmeDevice {
-                    model: "Micron_7450_MTFDKBG1T9TFR".into(),
-                    firmware_rev: "E2MU200".into(),
-                    serial: format!("MicronFAKESERNUM{n}"),
-                })
-                .chain((0..8).map(|n| NvmeDevice {
-                    model: "KCM6DRUL3T84".into(),
-                    firmware_rev: "0107".into(),
-                    serial: format!("KCMFAKESERNUM{n}"),
-                }))
-                .collect(),
-            dmi_data: Some(DmiData {
-                board_name: "DGXH100".into(),
-                board_version: "555.07L01.0001".into(),
-                bios_version: "1.6.7".into(),
-                bios_date: "02/20/2025".into(),
-                product_serial: self.dgx_system_serial_number.to_string(),
-                board_serial: format!("{}.FAKESERNUM1", self.dgx_system_serial_number),
-                chassis_serial: self.dgx_chassis_serial_number.to_string(),
-                product_name: "DGXH100".into(),
-                sys_vendor: "NVIDIA".into(),
-            }),
-            dpu_info: None,
-            gpus: (0..8)
-                .map(|n| {
-                    let pci_bus_id = [
-                        "00000000:1B:00.0",
-                        "00000000:43:00.0",
-                        "00000000:52:00.0",
-                        "00000000:61:00.0",
-                        "00000000:9D:00.0",
-                        "00000000:C3:00.0",
-                        "00000000:D1:00.0",
-                        "00000000:DF:00.0",
-                    ][n];
-                    Gpu {
-                        name: "NVIDIA H100 80GB HBM3".into(),
-                        serial: self.gpu_serial[n].to_string(),
-                        driver_version: "580.126.16".into(),
-                        vbios_version: "96.00.A5.00.01".into(),
-                        inforom_version: "G520.0200.00.05".into(),
-                        total_memory: "81559 MiB".into(),
-                        frequency: "1980 MHz".into(),
-                        pci_bus_id: pci_bus_id.into(),
-                        platform_info: None,
-                    }
-                })
-                .collect(),
-            memory_devices: (0..32)
-                .map(|_| MemoryDevice {
-                    size_mb: Some(65536),
-                    mem_type: Some("DDR5".into()),
-                })
-                .collect(),
-            tpm_ek_certificate: None,
-            tpm_description: None,
-            ..Default::default()
-        }
-    }
-
-    fn discovery_info_network_interfaces(&self) -> Vec<NetworkInterface> {
-        vec![
-            self.mgmt_nic.discovery_info(
-                "/devices/pci0000:00/0000:00:10.0/0000:0b:00.0/net/eno3",
-                "0000:0b:00.0",
-                0,
-            ),
-            self.storage_nic0.discovery_info(
-                0,
-                "/devices/pci0000:24\
-                /0000:24:01.0/0000:25:00.0/0000:26:00.0\
-                /0000:27:00.0/0000:28:00.0/0000:29:00.0\
-                /net/enp41s0f0np0",
-                "0000:29:00.0",
-                0,
-            ),
-            self.storage_nic0.discovery_info(
-                1,
-                "/devices/pci0000:24/0000:24:01.0\
-                /0000:25:00.0/0000:26:00.0/0000:27:00.0\
-                /0000:28:00.0/0000:29:00.1\
-                /net/enp41s0f1np1",
-                "0000:29:00.1",
-                0,
-            ),
-            self.dpu.host_nic_discovery_info(
-                "/devices/pci0000:80/0000:80:05.0/0000:82:00.0/net/ens6np0",
-                "0000:82:00.0",
-                0,
-            ),
-        ]
-    }
-
-    fn discovery_info_ib_interfaces(&self) -> Vec<InfinibandInterface> {
-        self.ib_nics
-            .iter()
-            .flat_map(|nic| nic.ib_nics())
-            .enumerate()
-            .map(|(n, _nic)| {
-                let (bus, numa_node) = [
-                    (0x15, 0),
-                    (0x3d, 0),
-                    (0x4c, 0),
-                    (0x5b, 0),
-                    (0x97, 1),
-                    (0xbd, 1),
-                    (0xcb, 1),
-                    (0xd9, 1),
-                ][n];
-                let device_name = format!("ibp{bus}s0");
-                let path = format!(
-                    "/devices/pci0000:{:02x}/0000:{:02x}:01.0/0000:{:02x}:00.0/\
-                     0000:{:02x}:00.0/0000:{:02x}:00.0/infiniband/{device_name}",
-                    bus,
-                    bus,
-                    bus + 1,
-                    bus + 2,
-                    bus + 3
-                );
-                InfinibandInterface {
-                    pci_properties: Some(PciDeviceProperties {
-                        vendor: "Mellanox Technologies".into(),
-                        device: "MT2910 Family [ConnectX-7]".into(),
-                        path,
-                        numa_node,
-                        description: Some("MT2910 Family [ConnectX-7]".into()),
-                        slot: format!("0000:{:02x}:00.0", bus + 3).into(),
-                    }),
-                    guid: format!("94dae0000000000{n}"),
-                }
-            })
-            .collect()
     }
 }
