+++ crates/bmc-mock/src/hw/nvidia_gb200.rs
@@ -18,9 +18,6 @@
 use std::borrow::Cow;
 use std::fmt;
 
-use rpc::PciDeviceProperties;
-use rpc::machine_discovery::{Gpu, GpuPlatformInfo, InfinibandInterface, MemoryDevice};
-
 use crate::redfish;
 
 #[derive(Clone, Copy)]
@@ -145,58 +142,9 @@ impl BiancaBoard<'_> {
             }
         })
     }
-
-    pub fn discovery_gpu(&self) -> [Gpu; 2] {
-        [0, 1].map(|gpun| Gpu {
-            name: "NVIDIA GB200".into(),
-            serial: self.gpu_serial_number.to_string(),
-            driver_version: "580.126.16".into(),
-            vbios_version: "97.00.B9.00.76".into(),
-            inforom_version: "G548.0201.00.06".into(),
-            total_memory: "189471 MiB".into(),
-            frequency: "2062 MHz".into(),
-            pci_bus_id: self.pcie_address(gpun).to_string(),
-            platform_info: Some(GpuPlatformInfo {
-                chassis_serial: format!("182100000000{}{gpun}", self.index),
-                slot_number: 24,
-                tray_index: 14,
-                host_id: 1,
-                module_id: self.module_id(gpun),
-                fabric_guid: format!("0xfeeeeeeeeeeeee{gpun:02x}"),
-            }),
-        })
-    }
-
-    pub fn discovery_memory(&self) -> MemoryDevice {
-        MemoryDevice {
-            size_mb: Some(491520),
-            mem_type: Some("LPDDR5".into()),
-        }
-    }
-
-    fn pcie_address(&self, gpu_index: u8) -> &'static str {
-        match (self.index, gpu_index) {
-            (BoardIndex::Board0, 0) => "00000008:01:00.0",
-            (BoardIndex::Board0, 1) => "00000009:01:00.0",
-            (BoardIndex::Board1, 0) => "00000018:01:00.0",
-            (BoardIndex::Board1, 1) => "00000019:01:00.0",
-            _ => panic!("unexpected gpu index: {gpu_index}"),
-        }
-    }
-
-    fn module_id(&self, gpu_index: u8) -> u32 {
-        match (self.index, gpu_index) {
-            (BoardIndex::Board0, 0) => 2,
-            (BoardIndex::Board0, 1) => 1,
-            (BoardIndex::Board1, 0) => 4,
-            (BoardIndex::Board1, 1) => 3,
-            _ => panic!("unexpected gpu index: {gpu_index}"),
-        }
-    }
 }
 
 pub struct IoBoard<'a> {
-    pub index: BoardIndex,
     pub serial_number: Cow<'a, str>,
 }
 
@@ -236,53 +184,6 @@ impl IoBoard<'_> {
             sensors: Some(sensors),
             id,
             ..redfish::chassis::SingleChassisConfig::defaults()
-        }
-    }
-
-    pub fn discovery_infiniband(&self) -> [InfinibandInterface; 2] {
-        [0, 1].map(|n| {
-            let numa_node = self.numa_node();
-            let domain = self.pcie_domain(n);
-            let device_name = if domain == 0 {
-                Cow::Borrowed("ibp3s0")
-            } else {
-                format!("ibP{domain}p3s0").into()
-            };
-            InfinibandInterface {
-                pci_properties: Some(PciDeviceProperties {
-                    vendor: "Mellanox Technologies".into(),
-                    device: "MT2910 Family [ConnectX-7]".into(),
-                    path: format!(
-                        "/devices/pci{domain:02x}:00\
-                             /{domain:02x}:00:00.0\
-                             /{domain:02x}:01:00.0\
-                             /{domain:02x}:02:00.0\
-                             /{domain:02x}:03:00.0\
-                             /infiniband/{device_name}"
-                    ),
-                    numa_node,
-                    description: Some("MT2910 Family [ConnectX-7]".into()),
-                    slot: format!("{domain}:03:00.0").into(),
-                }),
-                guid: format!("7c8c09000000000{n}"),
-            }
-        })
-    }
-
-    fn numa_node(&self) -> i32 {
-        match self.index {
-            BoardIndex::Board0 => 0,
-            BoardIndex::Board1 => 1,
-        }
-    }
-
-    fn pcie_domain(&self, dev_number: u8) -> u32 {
-        match (self.index, dev_number) {
-            (BoardIndex::Board0, 0) => 0x0000,
-            (BoardIndex::Board0, 1) => 0x0002,
-            (BoardIndex::Board1, 0) => 0x0010,
-            (BoardIndex::Board1, 1) => 0x0012,
-            _ => panic!("unexpected dev number: {dev_number}"),
         }
     }
 }
