+++ crates/bmc-mock/src/hw/supermicro_gb300_nvl.rs
@@ -25,7 +25,6 @@ use std::borrow::Cow;
 use std::sync::Arc;
 
 use mac_address::MacAddress;
-use rpc::DiscoveryInfo;
 use serde_json::json;
 
 use crate::{BootOptionKind, Callbacks, hw, redfish};
@@ -296,10 +295,6 @@ impl SupermicroGB300Nvl<'_> {
         redfish::update_service::UpdateServiceConfig {
             firmware_inventory: vec![],
         }
-    }
-
-    pub fn discovery_info(&self) -> DiscoveryInfo {
-        DiscoveryInfo::default()
     }
 }
